Add files via upload
To get this to work on your own machine, you need to make a Discord bot on the Discord Developer Portal and get its token. You will also need to download FFmpeg and place the files ffmpeg.exe, ffplay.exe, and ffprobe.exe in the same path folder as music-bot-v1.0.py (and then run it of course).